move to spark-core
